Latest updates and cleanup
OurWork tab showing pngs, not final, but starting to look/feel like I
want it to. App GUI tweaks all around, probably none of them “good
practice”, but I’m new to Swing. Just need to finish tackling the
layout view for the showcase video, and the home page (add Instagram)
and decide on the image layout and sizes, and then the actual video
players.